added batch action for overriding multiple (not working for community rowing inc unless punctuation is removed)
