+++ flagging_site/data/boathouses.py
@@ -3,13 +3,12 @@ from typing import Dict
 from typing import List
 
 from sqlalchemy import func
-from sqlalchemy.orm import Session
+
 from sqlalchemy.dialects.postgresql import aggregate_order_by
 
-from ..admin import ModelView
+
 from .database import db
 from .database import execute_sql
-# from .predictive_models import
 
 
 class Boathouse(db.Model):
@@ -131,20 +130,4 @@ def get_overridden_boathouses():
     return ret_val
 
 
-class ManualOverridesModelView(ModelView):
-    
-    form_choices = {
-        'reason': [
-            ('cyanobacteria', 'Cyanobacteria'),
-            ('sewage', 'Sewage'),
-            ('other', 'Other'),
-        ]
-    }
 
-    def __init__(self, session: Session):
-        super().__init__(
-            Boathouse,
-            session,
-            endpoint='manual_overrides',
-            name='Boathouses (including Manual Overrides)'
-        )
